Converted Video Games to Use New Sqlite Async Library
